requirements.txt
streamlit
langchain
langchain-google-genai
python-dotenv
nest-asyncio
streamlit-authenticator
PyYAML
bcrypt